Updated duplicate testing to use ginkgo
